Remove some old Go version specific logic
This fork doesn't aim at supporting very old Go versions, so the
logic that existed to keep the implementation backwards compatible
with those old Go versions was removed.
