+++ connector_test.go
@@ -1,5 +1,3 @@
-// +build go1.10
-
 package sqlmw
 
 import (
